Move event handling code into concern
